Make cert install idempotent

+++ nixpkgs/modules/desktop.nix
@@ -73,13 +73,16 @@
   };
 
   home.activation.installNssDbCerts = lib.hm.dag.entryAfter ["writeBoundary"] ''
-    ffdir=$HOME/.mozilla/firefox/$(ls $HOME/.mozilla/firefox | grep default)
-    if ! { [ -L "$ffdir/key4.db" ] && [ -L "$ffdir/cert9.db" ]; };
-    then 
-      $DRY_RUN_CMD mv $ffdir/key4.db $ffdir/cert9.db $HOME/.pki/nssdb/
-      $DRY_RUN_CMD ln -s ~/.pki/nssdb/key4.db $ffdir/key4.db
-      $DRY_RUN_CMD ln -s ~/.pki/nssdb/cert9.db $ffdir/cert9.db
+    crt="/etc/nixos/codexlab-ca.crt"
+    if [ -f "$crt" ]; then
+      ffdir=$HOME/.mozilla/firefox/$(ls $HOME/.mozilla/firefox | grep default)
+      if ! { [ -L "$ffdir/key4.db" ] && [ -L "$ffdir/cert9.db" ]; };
+      then 
+        $DRY_RUN_CMD mv $ffdir/key4.db $ffdir/cert9.db $HOME/.pki/nssdb/
+        $DRY_RUN_CMD ln -s ~/.pki/nssdb/key4.db $ffdir/key4.db
+        $DRY_RUN_CMD ln -s ~/.pki/nssdb/cert9.db $ffdir/cert9.db
+      fi
+      $DRY_RUN_CMD ${pkgs.nss.tools}/bin/certutil -A -t "C,," -n codexlab -i "$crt" -d sql:$HOME/.pki/nssdb
     fi
-    $DRY_RUN_CMD ${pkgs.nss.tools}/bin/certutil -A -t "C,," -n codexlab -i /etc/nixos/codexlab-ca.crt -d sql:$HOME/.pki/nssdb
   '';
 }
